backend/requirements.txt

flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0
kaggle==1.5.12
python-dotenv==0.19.1
gunicorn==20.1.0
